Add API tests
This commit refactors a small piece of the api
to make testing so much easier.

As it is very difficult and annoying to stub out
npm packages and the functions that come with them,
in order to easily test the api functions, we wrap
the firebase package functions in another variable,
which allows us to stub out the individual functions
for testing purposes.